% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromHMM.R
\name{bitmask.total}
\alias{bitmask.total}
\title{Compute number of covered bases}
\usage{
bitmask.total(bed)
}
\arguments{
\item{bed}{BED data.frame}
}
\value{
total number of covered bases
}
\description{
Compute number of covered bases
}
