% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{BwSet}
\alias{BwSet}
\title{Load the main datasets}
\usage{
BwSet(basePath = "/usr/data/", fileName)
}
\arguments{
\item{basePath}{Root directory for bigWig files}

\item{fileName}{Directory name for cell specific files}
}
\value{
List of bigWig objects
}
\description{
Load the main datasets
}
