% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromHMM.R
\name{chromHMM.regions}
\alias{chromHMM.regions}
\title{Load chromHMM regions for given cellLine}
\usage{
chromHMM.regions(
  cellLine,
  prom.states = "12",
  enh.states = "4567",
  basePath = "/usr/data/GROseq.parser/hg19/"
)
}
\arguments{
\item{cellLine}{cell line to use (gm12878 or k562)}

\item{prom.states}{character string with state numbers to use as promoters}

\item{enh.states}{character string with state numbers to use as enhancers}

\item{basePath}{Root directory for chromHMM bed files}
}
\value{
list with promoter(proms) and enhancer(enh) sets
}
\description{
Load chromHMM regions for given cellLine
}
