% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backcompat.R
\name{chromStepSum.bigWig}
\alias{chromStepSum.bigWig}
\title{deprecated function in bigWig library}
\usage{
chromStepSum.bigWig(bigwig, chrom, step, defaultValue)
}
\description{
deprecated function in bigWig library
}
