% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafuncs.R
\name{combine3.sequence.to.data}
\alias{combine3.sequence.to.data}
\title{Take a list of vectors of read counts (in 10bp steps) and produce a combined dataset of all processed vectors
needed by the HMM}
\usage{
combine3.sequence.to.data(
  plusGenestart,
  plusGenebody,
  plusGeneend,
  scale.factor,
  w = 10,
  k = 2,
  thresh = 1,
  testMode = FALSE
)
}
\arguments{
\item{plusGenestart}{list of read vectors}

\item{plusGenebody}{list of read vectors}

\item{plusGeneend}{list of read vectors}

\item{scale.factor}{scale factor}

\item{log2.thresh}{a step is peaked if value is above this value}
}
\value{
data suitable for the Naive HMMs
}
\description{
Take a list of vectors of read counts (in 10bp steps) and produce a combined dataset of all processed vectors
needed by the HMM
}
