% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normsize.R
\name{compute.normalization}
\alias{compute.normalization}
\title{Compute the total read count for each bigWig file in bwSet's list}
\usage{
compute.normalization(bwSet)
}
\arguments{
\item{bwSet}{A list of bigWig objects; for example, one produced by \code{mainBwSet.grocap}}
}
\value{
Single step sum of values across all chromosomes, for each bigWig file
}
\description{
Compute the total read count for each bigWig file in bwSet's list
}
