% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromHMM.R
\name{contained.fraction}
\alias{contained.fraction}
\title{compute fraction of regions contained within target set}
\usage{
contained.fraction(container, regions)
}
\arguments{
\item{container}{container regions}

\item{regions}{regions to test for inclusion in container regions}
}
\value{
fraction of source regions contained inside target set
}
\description{
compute fraction of regions contained within target set
}
