% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromHMM.R
\name{contained.mask}
\alias{contained.mask}
\title{identify regions contained within target set}
\usage{
contained.mask(container, regions)
}
\arguments{
\item{container}{container regions}

\item{regions}{regions to test for inclusion in container regions}
}
\value{
logical vector indicating inclusion of each region
}
\description{
identify regions contained within target set
}
