% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc.R
\name{create.single}
\alias{create.single}
\title{Select preds that have no neighbor up or down stream (up to thresh)}
\usage{
create.single(preds, thresh = 250)
}
\description{
Select preds that have no neighbor up or down stream (up to thresh)
}
