% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.parse.R
\name{decode.data.qhmm2Shapes}
\alias{decode.data.qhmm2Shapes}
\title{Decode HMM predictions}
\usage{
decode.data.qhmm2Shapes(
  hmm,
  data,
  start = 0,
  step = 50,
  TSS.range = NA,
  PG.S.range = NA,
  PG.B.range = NA,
  PG.E.range = NA,
  PAG.range = NA,
  PNEG.range = NA,
  PG.range = NA
)
}
\description{
Decode HMM predictions
}
