% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.parse.R
\name{decode.data.qhmmOld}
\alias{decode.data.qhmmOld}
\title{Decode HMM predictions}
\usage{
decode.data.qhmmOld(
  hmm,
  data,
  start = 0,
  step = 10,
  m1.range = 2:4,
  m2.range = NA,
  m2.start = 5,
  m2.mid = c(5, 6),
  m2.end = 7
)
}
\description{
Decode HMM predictions
}
