% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc.R
\name{edge.expand}
\alias{edge.expand}
\title{Edge extension}
\usage{
edge.expand(predictions, bwSet, bwBck, scale.factor, step = 10, thresh = 100)
}
\description{
edge extension: for predictions smaller than X bp (100 bp) find overlapping (up to) X bp window
                that maximally increases the number of stesp with TAP+ > TAP- (scaled).
                Don't extend if increase is zero; don't add useless edges (TAP- (scaled) > TAP+ or
                no data.), but these can be fixed on a subsequent edge-trim pass.
}
\details{
Secundary criterion to resolve ties: total read count (TAP+)
}
