% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{gcol}
\alias{gcol}
\title{Common color scheme}
\format{
An object of class \code{list} of length 5.
}
\usage{
gcol
}
\description{
\describe{
\item{GROcap}{green}
\item{CAGE}{orange}
\item{GROseq}{purple}
\item{Fwd}{red}
\item{Rev}{blue}
}
}
\keyword{datasets}
