% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.parse.R
\name{make.qhmm2Shapes11states11slots}
\alias{make.qhmm2Shapes11states11slots}
\title{Create HMM with 11 states and 11 observation slots (this one goes up to 11)}
\usage{
make.qhmm2Shapes11states11slots(
  tm = NULL,
  free.params.vec = c((1 - 0.9)/2, (1 - 0.9)/2, 0.9, 0.01, 0.45, 0.45)
)
}
\arguments{
\item{tm}{matrix of valid transitions}

\item{free.params.vec}{vector of free parameters}
}
\description{
Create HMM with 11 states and 11 observation slots (this one goes up to 11)
}
