% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.parse.R
\name{make.qhmm2Shapes4states3slots}
\alias{make.qhmm2Shapes4states3slots}
\title{Create HMM with 4 states and 3 observation slots}
\usage{
make.qhmm2Shapes4states3slots(
  free.params.vec = c((1 - 0.9)/2, (1 - 0.9)/2, 0.9, 0.01, 0.45, 0.45)
)
}
\description{
Create HMM with 4 states and 3 observation slots
}
