% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.parse.R
\name{make.qhmm2Shapes4states5slots}
\alias{make.qhmm2Shapes4states5slots}
\title{Create HMM with 4 states and 4 observation slots}
\usage{
make.qhmm2Shapes4states5slots(
  free.params.vec = c((1 - 0.9)/2, (1 - 0.9)/2, 0.9, 0.01, 0.45, 0.45)
)
}
\arguments{
\item{free.params.vec}{vector of free parameters}
}
\description{
Create HMM with 4 states and 4 observation slots
}
