% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromHMM.R
\name{overlaped.mask}
\alias{overlaped.mask}
\title{Identify regions overlapping within target set}
\usage{
overlaped.mask(setSrc, setSel)
}
\arguments{
\item{setSrc}{regions to process}

\item{setSel}{regions used to test overlap}
}
\value{
logical vector indicating overlap of each source region with selection set
}
\description{
Identify regions overlapping within target set
}
