% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromHMM.R
\name{pairs.in.regions}
\alias{pairs.in.regions}
\title{identify pairs contained within target set}
\usage{
\method{pairs}{in.regions}(pair.plus, pair.minus, container)
}
\arguments{
\item{pair.plus}{bed data.frame with plus side of pairs}

\item{pair.minus}{bed data.frame with minus side of pairs}

\item{container}{container regions}
}
\value{
logical vector indicating inclusion of each pair
}
\description{
identify pairs contained within target set
}
