% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.parse.R
\name{process.chromosome.qhmm}
\alias{process.chromosome.qhmm}
\title{Process chromosome with HMM}
\usage{
process.chromosome.qhmm(
  bwSet,
  chrom,
  scale.factor,
  step = 50,
  hmm = make.qhmm2Shapes2states5slots(),
  covar.lst = NULL,
  enable.EM = FALSE,
  thresh = 1,
  free.param.lst = NULL,
  testMode = FALSE
)
}
\description{
Process chromosome with HMM
}
