% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.parse.R
\name{process.genome.qhmm}
\alias{process.genome.qhmm}
\title{Process entire genome with HMM}
\usage{
process.genome.qhmm(
  bwSet,
  chroms,
  scale.factor,
  step = 50,
  hmm = make.qhmm2Shapes2states5slots(),
  covar.lst = NULL,
  enable.EM = FALSE,
  thresh = 1,
  free.param.lst = NULL,
  testMode = FALSE
)
}
\description{
Process entire genome with HMM
}
