% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafuncs.R
\name{processReads.globalMean}
\alias{processReads.globalMean}
\title{Take a vector of read values (in 10bp steps) and produce a thresholded vector of zeros and ones}
\usage{
processReads.globalMean(reads, scale.factor, thresh, testMode = FALSE)
}
\arguments{
\item{reads}{list of read vectors}

\item{scale.factor}{scale factor}

\item{thresh}{a step is peaked if value is above this value}
}
\value{
data suitable for the Naive HMMs
}
\description{
Take a vector of read values (in 10bp steps) and produce a thresholded vector of zeros and ones
}
