% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafuncs.R
\name{processReads}
\alias{processReads}
\title{Take a vector of read values (in 10bp steps) and produce a thresholded vector of zeros and ones}
\usage{
processReads(reads, scale.factor, thresh, testMode = FALSE, w = 10, k = 2)
}
\arguments{
\item{reads}{list of read vectors}

\item{scale.factor}{scale factor}

\item{thresh}{a step is peaked if value is above this value}

\item{w}{number of steps to look for background values (pos +/- w)}

\item{k}{minimum number of background steps to run log2 ratio test}
}
\value{
data suitable for the Naive HMMs
}
\description{
Take a vector of read values (in 10bp steps) and produce a thresholded vector of zeros and ones
}
