% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backcompat.R
\name{queryByStep.bigWig}
\alias{queryByStep.bigWig}
\title{deprecated function in bigWig library}
\usage{
queryByStep.bigWig(
  bigWig,
  chrom,
  start,
  end,
  step,
  do.sum = FALSE,
  default.null = TRUE,
  defaultValue = 0
)
}
\description{
deprecated function in bigWig library
}
