% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafuncs.R
\name{sequence.to.data}
\alias{sequence.to.data}
\title{Take a vector of read counts (in 10bp steps) and produce the (Y,X) pair
needed by the HMM}
\usage{
sequence.to.data(
  reads.in.steps.tap,
  reads.in.steps.notap,
  scale.factor,
  w = 10,
  k = 2,
  log2.thresh = 1
)
}
\arguments{
\item{reads.in.steps.tap}{reads in TAP+ condition}

\item{reads.in.steps.notap}{reads in TAP- condition}

\item{scale.factor}{TAP- to TAP+ scale factor}

\item{w}{number of steps to look for background values (pos +/- w)}

\item{k}{minimum number of background steps to run log2 ratio test}

\item{log2.thresh}{a step is peaked if ratio to background windows is above this value}
}
\value{
data suitable for the Naive HMMs
}
\description{
Peaks identification makes use of steps in the local region marked as depleated. If a sufficient number is present, a step is called a peak if the TAP+ read count is above the mean TAP+ reads in neighboring background (depleated) steps.
}
