% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{stable.unstable.classify}
\alias{stable.unstable.classify}
\title{Classify pairs (in order, [-, +]) as SS, SU, US, UU}
\usage{
stable.unstable.classify(pair.plus, pair.minus, bwSet, cage.thresh = 8)
}
\description{
Classify pairs (in order, [-, +]) as SS, SU, US, UU
}
