% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc.stats.R
\name{step.coverage}
\alias{step.coverage}
\title{step type coverage}
\usage{
step.coverage(
  preds,
  chroms,
  bwSet,
  bwBck,
  scale.factor,
  step = 10,
  n.cores = 1
)
}
\description{
step type coverage
}
