% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss.pairs.R
\name{values.preds}
\alias{values.preds}
\title{Collect GROcap and CAGE (or other bigWig dataset) in a set of HMM TSS predictions}
\usage{
values.preds(preds, bwGROcap, bwCAGE, hwin = 50, snd.win = 0)
}
\arguments{
\item{preds}{BED data-frame with HMM identified TSS regions}

\item{bwGROcap}{bigWig object with GRO-cap data (in the strand matching 'preds')}

\item{bwCAGE}{bigWig object with CAGE (or other) data (in the strand matching 'preds')}

\item{hwin}{half-window in bp (see details)}

\item{snd.win}{extension applied to the collection window for the second dataset (bwCAGE). Window is extended in the strand sense.}
}
\value{
Nx2 matrix with data for each prediction
}
\description{
If predictions are larger than 2*hwin bp, then their GRO-cap weighted center
is computed and a 2*hwin window around that center is used.
}
