% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafuncs.R
\name{write.track.bed}
\alias{write.track.bed}
\title{Write a BED data frame as a prediction track (with colors)}
\usage{
write.track.bed(
  bed,
  filename,
  track = "hmm.preds",
  color.plus = "197,0,11",
  color.minus = "0,132,209"
)
}
\arguments{
\item{bed}{BED data frame}

\item{filename}{output BED filename}

\item{track}{track/description string}

\item{color.plus}{RGB color (comma separated) for plus strand}

\item{color.minus}{RGB color (comma separated) for minus strand}
}
\description{
Write a BED data frame as a prediction track (with colors)
}
